examples/lab5/requirements.txt
datasets==1.6.2
scikit-learn==0.24.1
PyYAML==5.4.1
dvc==2.0.18
